// src/app/dto/find-streams-response.dto.ts
export interface FindStreamsResponseDto {
  data: FindStreamData[];
}

// export interface StreamData {
export interface FindStreamData {
  id: string;
  user_id: string;
  user_login: string;
  user_name: string;
  game_id: string;
  game_name: string;
  type: string;
  title: string;
  viewer_count: number;
  started_at: string;
  language: string;
  thumbnail_url: string;
  tag_ids: string[];
  is_mature: boolean;
}